pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools_scm]
version_file = "softorks/softorks/__init__.py"

[project]
name = "softorks.eu"
description="WWW website of mine - softorks.eu and softarna.cz"
dynamic = ["version"]
authors = [
    {name = "Michal Mládek"},
    {email = "michal.mladek.78@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
keywords = ["website", "internet"]
classifiers = [
    "Framework :: Django",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv",
    "django",
    "django-extensions",
    "django-modeltranslation",
    "markdown",
    "pydantic",
]

[project.license]
file = "LICENSE"

[project.optional-dependencies]
dev = [
    "httpx",
    "rich",
    "ipython",
    "notebook",
    "build",
    "validate-pyproject",
    "pre-commit",
    "black",
    "autoflake",
    "flake8",
    "mypy",
    "django-stubs[compatible-mypy]",
    "django-stubs-ext",
    "types-markdown",
    "isort",
]
prod = ["gunicorn[gevent]"]

[project.urls]
homepage = "https://softorks.eu"
repository = "https://github.com/michalnik/softorks.eu"
